+++ src/experiments/08/Experiment08.cpp
@@ -59,7 +59,7 @@ namespace waveletExperiments
 			/**
 			 * Wavelet waveform function
 			 */
-			static inline std::vector<double> wavelet;
+			static inline std::vector<long double> wavelet;
 
 		public:
 
@@ -81,7 +81,7 @@ namespace waveletExperiments
 			 * @param samplingRate
 			 * @param path
 			 */
-			static void waveletAnaliticFunction(std::vector<double> &signal, int &signalLength, unsigned int samplingRate, std::string path)
+			static void waveletAnaliticFunction(std::vector<long double> &signal, int &signalLength, unsigned int samplingRate, std::string path)
 			{
 
 				///////////////////////
@@ -111,7 +111,7 @@ namespace waveletExperiments
 
 				// features vector has the amount of values equals to amount of the ranges minus 1 
 				// because we are summing up intervals
-				std::vector<double> featureVector(melRanges.size() - 1);
+				std::vector<long double> featureVector(melRanges.size() - 1);
 
 				// We need to known the max frequency supported
 				// by the signal in order to find the values in
@@ -145,7 +145,7 @@ namespace waveletExperiments
 					{
 
 						// Retrieve the values
-						std::vector<double> sig1 = transformedSignal.getWaveletPacketTransforms(startIndex);
+						std::vector<long double> sig1 = transformedSignal.getWaveletPacketTransforms(startIndex);
 
 						// Sum the power of 2 of them all!!! (i.e. calculate the energies)
 						featureVector.at(i) = 0;
@@ -237,7 +237,7 @@ namespace waveletExperiments
 				 * 			etc.
 				 * 		Etc.
 				 */
-				std::map<std::string, std::map<BARK_MEL, std::map<std::string, std::vector<std::vector<double>>>>> results;
+				std::map<std::string, std::map<BARK_MEL, std::map<std::string, std::vector<std::vector<long double>>>>> results;
 
 				////////////////////////////////////////
 				/// Preparing to compute the progress //
@@ -312,16 +312,16 @@ namespace waveletExperiments
 				std::map<double, std::vector<statistics::ConfusionMatrix>> confusionMatricesForEachPercentage;
 
 				// Holds the tests features vectors for live signals
-				std::vector<std::vector<double>> testLive;
+				std::vector<std::vector<long double>> testLive;
 
 				// Holds the models features vectors for live signals
-				std::vector<std::vector<double>> modelLive;
+				std::vector<std::vector<long double>> modelLive;
 
 				// Holds the tests features vectors for spoofed signals
-				std::vector<std::vector<double>> testSpoofing;
+				std::vector<std::vector<long double>> testSpoofing;
 
 				// Holds the models features vectors for spoofed signals
-				std::vector<std::vector<double>> modelSpoofing;
+				std::vector<std::vector<long double>> modelSpoofing;
 
 				// Creating the classifier
 				classifiers::SupportVectorMachine c;
